+++ kevlar/__init__.py
@@ -51,10 +51,10 @@ from kevlar import partition
 from kevlar import localize
 from kevlar import call
 from kevlar import alac
-from kevlar import simplex
 from kevlar import simlike
 from kevlar import dist
 from kevlar import split
+from kevlar import unband
 from kevlar import gentrio
 from kevlar import cli